requirements.txt
# HelixForge - Cross-Dataset Insight Synthesizer
# Python Requirements

# Core
pandas>=2.0
numpy>=1.24
pyarrow>=14.0

# API
fastapi>=0.109
uvicorn>=0.27
pydantic>=2.5

# LLM / Embeddings
openai>=1.0
tiktoken>=0.5

# ML / Stats
scikit-learn>=1.3
scipy>=1.11

# Utilities
chardet>=5.0
fuzzywuzzy>=0.18
python-Levenshtein>=0.21
pyyaml>=6.0
openpyxl>=3.1

# Testing
pytest>=7.4
pytest-asyncio>=0.23
httpx>=0.26
pytest-cov>=4.1
